% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_imports.R, R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{get_raw_data Gets raw nominal house price data from LU Open Data Portal}
\usage{
get_raw_data(
  url = "https://data.public.lu/fr/datasets/r/1d20f982-57e1-4ae2-a278-dc78c88c21dc"
)

get_raw_data(
  url = "https://data.public.lu/fr/datasets/r/1d20f982-57e1-4ae2-a278-dc78c88c21dc"
)
}
\arguments{
\item{url}{Optional: Persistent url to the data}
}
\value{
A data frame

A data frame
}
\description{
get_raw_data Gets raw nominal house price data from LU Open Data Portal

get_raw_data Gets raw nominal house price data from LU Open Data Portal
}
