% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luHousePrices-package.R
\docType{package}
\name{luHousePrices-package}
\alias{luHousePrices}
\alias{luHousePrices-package}
\title{define_imports This function lists the needed dependencies}
\value{
Nothing; simply used for NAMESPACE creation
}
\description{
Contains house prices from the 'L’Observatoire de l’Habitat'.
}
\author{
\strong{Maintainer}: Bruno Rodrigues \email{bruno@brodrigues.co} (\href{https://orcid.org/0000-0002-3211-3689}{ORCID})

Other contributors:
\itemize{
  \item MESR [copyright holder]
}

}
\keyword{internal}
