% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_imports.R, R/get_raw_data.R
\name{make_commune_level_data}
\alias{make_commune_level_data}
\title{make_commune_level_data Makes the final data at commune level}
\usage{
make_commune_level_data(flat_data)

make_commune_level_data(flat_data)
}
\arguments{
\item{flat_data}{Flat data df as returned by clean_flat_data()}
}
\value{
A data frame

A data frame
}
\description{
make_commune_level_data Makes the final data at commune level

make_commune_level_data Makes the final data at commune level
}
