% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_imports.R, R/get_raw_data.R
\name{clean_raw_data}
\alias{clean_raw_data}
\title{clean_raw_data Cleans the raw data}
\usage{
clean_raw_data(raw_data)

clean_raw_data(raw_data)
}
\arguments{
\item{url}{Optional: Persistent url to the data}
}
\value{
A data frame

A data frame
}
\description{
clean_raw_data Cleans the raw data

clean_raw_data Cleans the raw data
}
\details{
Removes uneeded rows, renames localities to make their names consistent across years and converts columns with prices to numeric columns

Removes uneeded rows, renames localities to make their names consistent across years and converts columns with prices to numeric columns
}
